refactor(scripts): simplify render function
