+++ package.json
@@ -108,7 +108,6 @@
     "jest-axe": "^3.1.1",
     "jest-cli": "^24.5.0",
     "jest-emotion": "^10.0.10",
-    "jsdom": "^15.1.1",
     "license-checker": "^25.0.1",
     "onchange": "^5.2.0",
     "postcss-loader": "^3.0.0",
